implemented JSON config loading, god this took forever

+++ source/json.h
@@ -1,7 +1,7 @@
 #ifndef JSON_H
 #define JSON_H
 
-int jsonLoad(char* path, int selCount, char** selections, int* selectArray);
-int jsonSave(char* path, int selCount, char** selections, int* selectArray);
+int jsonLoad(char* path, s16 selCount, char** selections, u8* selectArray);
+int jsonSave(char* path, int selCount, char** selections, u8* selectArray);
 
 #endif
